Review feedback and component provisioning
Shift to functional components
Abstract both list and list item away for re-use
Roll unit test result specific styling into UT source
Roll supporting UI over including badges > extend SC for flavors
Shift timestamp to render the value, JSX handling formatting
