+++ packages/insomnia-app/app/ui/css/components/unit-tests.less
@@ -71,11 +71,7 @@
     overflow-y: auto;
 
     ul {
-      padding-top: var(--padding-md);
-      li {
-        display: block;
-        padding: var(--padding-xs) var(--padding-xs) var(--padding-xs) var(--padding-xxs);
-      }
+      padding-top: var(--padding-sm);
     }
   }
