+++ packages/insomnia-components/components/list-group.stories.js
@@ -7,20 +7,9 @@ export default { title: 'Lists | List Group' };
 export const _default = () => (
   <div style={{ width: '400px' }}>
     <ListGroup>
-      <ListGroupItem
-        name="My test name one"
-        time="602ms"
-        result="fail"
-        errorMsg="Some dastardly error."
-      />
-      <ListGroupItem name="My test name two" time="6ms" result="pass" />
-      <ListGroupItem name="My test name three" time="22ms" result="pass" />
-      <ListGroupItem
-        name="My test name four"
-        time="602ms"
-        result="fail"
-        errorMsg="Some dastardly verbose error consuming a lot of space."
-      />
+      <ListGroupItem>List</ListGroupItem>
+      <ListGroupItem>of</ListGroupItem>
+      <ListGroupItem>things...</ListGroupItem>
     </ListGroup>
   </div>
 );
